GP compatibility (#27)
Changes:
- Make it possible to compile "bare" service without SDK dependencies
- Add bare-fibonacci

This make it possible for the playground to provide GP conformant
Service blobs. In the future we might want to fork Polkatool
and change the encoding to be directly GP conformant. Currently I only
fork Polkatool to make some build dependencies optional
that were failing to build in docker.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver@tasty.limo>